server/requirements.txt
aiofiles==0.7.0
anyio==3.3.0
bidict==0.21.3
certifi==2021.5.30
charset-normalizer==2.0.4
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httpcore==0.13.6
httptools==0.3.0
httpx==0.19.0
hyperframe==6.0.1
idna==3.2
multidict==5.1.0
pytchat==0.5.5
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
rfc3986==1.5.0
sanic==21.6.2
sanic-routing==0.7.1
sniffio==1.2.0
urllib3==1.26.6
websockets==10.0
